+++ CHANGELOG.md
@@ -77,6 +77,25 @@ the following form: YYYY.0M.0D.
 - Created initial drafts of the six portfolio assessments
 - Added READMEs to key folders like `test` and `lib` to explain their purpose
 
+## [2026.04.15]
+
+### Added
+
+- Created class CalorieTracker1 extending CalorieTrackerSecondary
+- Implemented kernel methods 'addRecord', 'removeRecord', and 'totalCaloriesOn'
+- Added constructor and Standard methods 'newInstance', 'clear', and 'transferFrom'
+
+
+## 2026.04.24
+
+
+### Added
+- Added kernel methods 'hasEntry' and 'entryCount'
+
+### Updated
+- Revised CalorieTracker method names
+- Revised assertion error messages
+
 [unreleased]: https://github.com/jrg94/portfolio-project/compare/v2024.08.07...HEAD
 [2024.08.07]: https://github.com/jrg94/portfolio-project/compare/v2024.01.07...v2024.08.07
 [2024.01.07]: https://github.com/jrg94/portfolio-project/releases/tag/v2024.01.07